Accessibility and HTML on the index
Added accessibility and changed a little HTML on the front page, also slightly indented the code.
Transparent in terms of rendering, better for SEO, better for people with disabilities.
Updated styles a little to add contrast and affordance
Added some OpQuast-compatible semantics
Added "by" on the homepage footer
Added simple tests on Twig for loops
Reformatted code to have the same standards in all file types